fixed quantity to float

+++ be/app/models/item/model.py
@@ -16,7 +16,7 @@ item_types = ["component", "ingredient", "tool", "equipment"]
 class ItemBase(SQLModel):
     name: str = Field(index=True)
     item_type: str = Field(...)
-    quantity: int = Field(...)
+    quantity: float = Field(...)
     unit: str = Field(...)
     pocket_id: int = Field(foreign_key="pocket.id")
     chest_id: int = Field(foreign_key="chest.id")
